[MAJOR] EDA-595, EDA-597 Create a view and a template for edit supplier's info and for approval workflow
